WIP Saturate DBusMessage caching in libdbus to help memory leak analysis
The libdbus keeps a cache DBusMessage objects and recycles them. While
this is good from performance point of view, it also means that some
messages are never released and are attributed in valgrind logs to the
place that happened to do the initial allocation. This makes it difficult
to see if messages are leaked or not.

By creating a number of dummy dbus messages and then releasing them
we can saturate the caching mechanism. After this we can be relatively
sure that DBusMessages are not being leaked as long all of them are
reported to originate from cache stuffing function.